backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
pydantic==2.10.0
pydantic-settings==2.6.0

# RSS parsing
feedparser==6.0.11

# HTTP client
httpx==0.26.0

# Audio conversion
ffmpeg-python==0.2.0

# Task scheduling
apscheduler==3.10.4

# WebSocket
websockets==12.0

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1
